feat(web): lift the routes to app
