+++ web/src/features/panels/LeftPanel.tsx
@@ -5,70 +5,13 @@ import { useAtom } from 'jotai';
 import { ChevronLeft, ChevronRight } from 'lucide-react';
 import { lazy, Suspense } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  Navigate,
-  Route,
-  Routes,
-  useLocation,
-  useParams,
-  useSearchParams,
-} from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 import { useIsMobile } from 'utils/styling';
 
 import { leftPanelOpenAtom } from './panelAtoms';
-import { zoneExists } from './zone/util';
 
 const RankingPanel = lazy(() => import('./ranking-panel/RankingPanel'));
 const ZoneDetails = lazy(() => import('./zone/ZoneDetails'));
-
-function HandleLegacyRoutes() {
-  const [searchParameters] = useSearchParams();
-
-  const page = (searchParameters.get('page') || 'map')
-    .replace('country', 'zone')
-    .replace('highscore', 'ranking');
-  searchParameters.delete('page');
-
-  const zoneId = searchParameters.get('countryCode');
-  searchParameters.delete('countryCode');
-
-  return (
-    <Navigate
-      to={{
-        pathname: zoneId ? `/zone/${zoneId}` : `/${page}`,
-        search: searchParameters.toString(),
-      }}
-    />
-  );
-}
-
-function ValidZoneIdGuardWrapper({ children }: { children: JSX.Element }) {
-  const [searchParameters] = useSearchParams();
-  const { zoneId } = useParams();
-
-  if (!zoneId) {
-    return <Navigate to="/" replace />;
-  }
-  const upperCaseZoneId = zoneId.toUpperCase();
-  if (zoneId !== upperCaseZoneId) {
-    return <Navigate to={`/zone/${upperCaseZoneId}?${searchParameters}`} replace />;
-  }
-
-  // Handle legacy Australia zone names
-  if (upperCaseZoneId.startsWith('AUS')) {
-    return (
-      <Navigate to={`/zone/${zoneId.replace('AUS', 'AU')}?${searchParameters}`} replace />
-    );
-  }
-
-  // Only allow valid zone ids
-  // TODO: This should redirect to a 404 page specifically for zones
-  if (!zoneExists(upperCaseZoneId)) {
-    return <Navigate to="/" replace />;
-  }
-
-  return children;
-}
 
 type CollapseButtonProps = {
   isCollapsed: boolean;
@@ -102,10 +45,11 @@ function MobileHeader() {
   );
 }
 
-function OuterPanel({ children }: { children: React.ReactNode }) {
+export default function LeftPanel() {
   const [isOpen, setOpen] = useAtom(leftPanelOpenAtom);
   const location = useLocation();
   const isMobile = useIsMobile();
+  const isZonePath = location.pathname.startsWith('/zone/');
 
   const onCollapse = () => setOpen(!isOpen);
 
@@ -117,37 +61,12 @@ function OuterPanel({ children }: { children: React.ReactNode }) {
       } ${isOpen ? '' : '-translate-x-full'}`}
     >
       {isMobile && <MobileHeader />}
-      <section className="h-full w-full">{children}</section>
+      <section className="h-full w-full">
+        <Suspense fallback={<LoadingSpinner />}>
+          {isZonePath ? <ZoneDetails /> : <RankingPanel />}
+        </Suspense>
+      </section>
       <CollapseButton isCollapsed={!isOpen} onCollapse={onCollapse} />
     </div>
-  );
-}
-export default function LeftPanel() {
-  return (
-    <OuterPanel>
-      <Routes>
-        <Route path="/" element={<HandleLegacyRoutes />} />
-        <Route path="/zone" element={<Navigate to="/" replace />} />
-        <Route
-          path="/zone/:zoneId"
-          element={
-            <ValidZoneIdGuardWrapper>
-              <Suspense fallback={<LoadingSpinner />}>
-                <ZoneDetails />
-              </Suspense>
-            </ValidZoneIdGuardWrapper>
-          }
-        />
-        {/* Alternative: add /map here and have a NotFound component for anything else*/}
-        <Route
-          path="*"
-          element={
-            <Suspense fallback={<LoadingSpinner />}>
-              <RankingPanel />
-            </Suspense>
-          }
-        />
-      </Routes>
-    </OuterPanel>
   );
 }
